[android] Share common code between test runners
