+++ platform/android/src/test/collator_test_stub.cpp
