+++ platform/android/src/test/number_format_test_stub.cpp
